Move common functionality from Feed to Entity
